Configuraciones en el proyecto para refactorizar:
Configuracion para la DB
Creacion de modulo modelos
Ubicacion de los endpoints en un unico modulo
Registrar blueprint del CRUD de notas en app
