+++ flaskr/app.py
@@ -1,33 +1,19 @@
-import os
-from flask_sqlalchemy import SQLAlchemy
-from flask import Flask, redirect, request, jsonify, render_template, url_for
+from flaskr.config import Config
+from flaskr.models import db, Note
+from flaskr.notes.routes import notes_bp
+from flask import Flask, request, render_template
 
-DB_FILE = os.path.join(os.path.dirname(__file__), 'database.sqlite')
 
 app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///{DB_FILE}'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False 
+app.config.from_object(Config)
+db.init_app(app)
+app.register_blueprint(notes_bp)
 
-db = SQLAlchemy(app)
-
-
-class Note(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    title = db.Column(db.String(100), nullable=False)
-    content = db.Column(db.Text, nullable=False)
-
-    def __repr__(self):
-        return f'<Note {self.id}:{self.title}>'
-
-
-@app.route('/')
-def home():
-    notes = Note.query.all()
-    return render_template('home.html', notes=notes)
 
 @app.route('/acerca-de')
 def about():
     return "Esta es una aplicacion de prueba con flask."
+
 
 @app.route('/contacto', methods=['GET', 'POST'])
 def contact():
@@ -36,52 +22,8 @@ def contact():
     return "Pagina de contacto. Puedes enviar un mensaje aquí."
 
 
-@app.route('/api/info', methods=['GET'])
-def api_info():
-    return jsonify({
-        "name": "Flask App",
-        "version": "1.0.0",
-        "description": "A simple"
-    }), 200
+@app.route('/')
+def home():
+    notes = Note.query.all()
+    return render_template('home.html', notes=notes)
 
-
-@app.route('/confirmacion')
-def confirmation():
-    return "Confirmación de acción exitosa.", 200
-
-@app.route('/crear-nota', methods=['GET', 'POST'])
-def create_note():
-    if request.method == 'POST':
-        title = request.form.get('title', "")
-        content = request.form.get('content', "")
-
-        note_db = Note(title=title, content=content)
-
-        db.session.add(note_db)
-        db.session.commit()
-
-        return redirect(url_for('home'))
-    return render_template('note_form.html')
-
-
-@app.route('/editar-nota/<int:id>', methods=['GET', 'POST'])
-def edit_note(id: int):
-    note = Note.query.get_or_404(id)
-    if request.method == 'POST':
-        title = request.form.get('title', "")
-        content = request.form.get('content', "")
-
-        note.title = title
-        note.content = content
-
-        db.session.commit()
-        return redirect(url_for('home'))
-    return render_template('edit_note_form.html', note=note)
-
-
-@app.route('/eliminar-nota/<int:id>', methods=['POST'])
-def delete_note(id: int):
-    note = Note.query.get_or_404(id)
-    db.session.delete(note)
-    db.session.commit()
-    return redirect(url_for('home'))
