+++ flaskr/config.py
@@ -1,0 +1,12 @@
+import os
+
+
+DB_FILE = os.path.join(os.path.dirname(__file__), 'database.sqlite')
+
+class Config:
+    SQLALCHEMY_DATABASE_URI = f'sqlite:///{DB_FILE}'
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+    # @staticmethod
+    # def init_app(app):
+    #     pass  # Additional initialization can be done here if needed
