+++ flaskr/templates/home.html
@@ -15,15 +15,16 @@
                 {{ note.title }}
                 <br>
                 {{ note.content }}
-                <a href="{{ url_for('edit_note', id=note.id) }}">Editar Nota</a>
-                <form action="{{ url_for('delete_note', id=note.id) }}" method="post">
+                <a href="{{ url_for('notes.edit_note', id=note.id) }}">Editar Nota</a>
+                <br>
+                <form action="{{ url_for('notes.delete_note', id=note.id) }}" method="post">
                     <button type="submit">Eliminar Nota</button>
                 </form>
             </div>
         </li>
         {% endfor %}
     </ul>
-    <a href="{{ url_for('create_note') }}">agregar una nueva nota</a>
+    <a href="{{ url_for('notes.create_note') }}">agregar una nueva nota</a>
     
 </body>
 </html>
